Create Hangman
This game will take input from user and will see if the letter or letters matches to the selected word from the list. If user guess the entire word then it gives separate message and ends program. User will loop until they decide not to which they can press F to exit.